开发1024点一致性hash映射

+++ smart-jdbc/smart-jdbc-demo/src/main/java/com/yryz/smart/jdbc/demo/dao/persistence/OrderDao.java
@@ -28,7 +28,7 @@ public interface OrderDao {
 
     void truncateTable();
 
-    Long insert(Order model);
+    int insert(Order model);
 
     void delete(Long orderId);
